Add html and css to the repo
This is the 1st version of git commit.In this
version I have completed the 1st header part and
still have to complete the rest of the website.
I am commiting this because I'm going to make some
changes which could change the whole layout and
make my task at hand more difficult.So this commit
will help me recover my original header code.